+++ .github/CHANGELOG.md
@@ -175,6 +175,7 @@
 1. [MCDU] Improved mcdu server command line handling  - @frankkopp (Cdr_Maverick#6475)
 1. [RA] Add dual Radio Altimeter simulation - @beheh (Benedict Etzel)
 1. [MCDU] Improved Remote MCDU communication by reducing update frequency - @frankkopp (Frank Kopp)
+1. [FMGC] Re-implement altitude and speed constraints - @tracernz (Mike)
 
 ## 0.7.0